Removing lua_utils from storage_utils
