+++ scripts/callbacks/hourly/system/disk_monitor.lua
@@ -4,8 +4,11 @@
 
 local dirs = ntop.getDirs()
 package.path = dirs.installdir .. "/scripts/lua/modules/?.lua;" .. package.path
+require "profiling"
+profiling.get_current_memory("disk_monitor.lua")
 
 local storage_utils = require("storage_utils")
+profiling.get_current_memory("disk_monitor.lua")
 
 -- ##############################################
