+++ scripts/lua/alert_stats.lua
@@ -5,7 +5,7 @@ local dirs = ntop.getDirs()
 package.path = dirs.installdir .. "/scripts/lua/modules/?.lua;" .. package.path
 package.path = dirs.installdir .. "/scripts/lua/modules/alert_store/?.lua;" .. package.path
 
-require "lua_utils"
+require "lua_utils_generic"
 local page_utils = require "page_utils"
 local os_utils = require "os_utils"
 local ui_utils = require "ui_utils"
