+++ scripts/lua/flow_details.lua
@@ -4,9 +4,9 @@
 local dirs = ntop.getDirs()
 package.path = dirs.installdir .. "/scripts/lua/modules/?.lua;" .. package.path
 
-require "lua_utils"
 require "flow_utils"
 require "voip_utils"
+require "label_utils"
 
 local shaper_utils
