+++ scripts/lua/modules/historical_flow_utils.lua
@@ -2,7 +2,7 @@
 -- (C) 2013-24 - ntop.org
 --
 
-
+require "lua_utils_generic"
 local tag_utils = require "tag_utils"
 local dscp_consts = require "dscp_consts"
 local alert_consts = require "alert_consts"
